Fix non file read DayTimeInterval errors
