+++ integration_tests/src/main/python/arithmetic_ops_test.py
@@ -1027,14 +1027,14 @@ def test_abs_ansi_no_overflow_day_time_interval(ansi_enabled):
 @pytest.mark.parametrize('ansi_enabled', ['false', 'true'])
 def test_abs_ansi_overflow_day_time_interval(ansi_enabled):
     """
-    We don't check the error messages because they are different on CPU and GPU.
+    Check the error message only when ANSI mode is false because they are different on CPU and GPU.
     CPU: long overflow.
     GPU: One or more rows overflow for abs operation.
     """
     assert_gpu_and_cpu_error(
         df_fun=lambda spark: _get_overflow_df(spark, [timedelta(microseconds=LONG_MIN)], DayTimeIntervalType(), 'abs(a)').collect(),
         conf={'spark.sql.ansi.enabled': ansi_enabled},
-        error_message='SparkArithmeticException')
+        error_message='' if ansi_enabled else 'SparkArithmeticException')
 
 @pytest.mark.skipif(is_before_spark_330(), reason='DayTimeInterval is not supported before Pyspark 3.3.0')
 @pytest.mark.parametrize('ansi_enabled', ['false', 'true'])
