+++ sql-plugin/src/main/330+-nondb/scala/com/nvidia/spark/rapids/shims/Spark330PlusShims.scala
@@ -29,12 +29,10 @@ import org.apache.spark.sql.execution.datasources.{DataSourceUtils, FilePartitio
 import org.apache.spark.sql.execution.datasources.parquet.ParquetFilters
 import org.apache.spark.sql.execution.datasources.v2.BatchScanExec
 import org.apache.spark.sql.execution.python.PythonMapInArrowExec
-import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.rapids._
 import org.apache.spark.sql.rapids.execution.python.GpuPythonMapInArrowExecMeta
-import org.apache.spark.sql.rapids.shims.{GpuDivideDTInterval, GpuDivideYMInterval, GpuMultiplyDTInterval, GpuMultiplyYMInterval, GpuTimeAdd}
-import org.apache.spark.sql.types.{CalendarIntervalType, DayTimeIntervalType, DecimalType, StructType}
-import org.apache.spark.unsafe.types.CalendarInterval
+import org.apache.spark.sql.rapids.shims.{GpuDivideYMInterval, GpuMultiplyYMInterval}
+import org.apache.spark.sql.types.{DecimalType, StructType}
 
 trait Spark330PlusShims extends Spark321PlusShims with Spark320PlusNonDBShims {
 
@@ -144,49 +142,6 @@ trait Spark330PlusShims extends Spark321PlusShims with Spark320PlusNonDBShims {
             GpuFloor(lhs, floor.dataType)
           }
         }),
-      GpuOverrides.expr[TimeAdd](
-        "Adds interval to timestamp",
-        ExprChecks.binaryProject(TypeSig.TIMESTAMP, TypeSig.TIMESTAMP,
-          ("start", TypeSig.TIMESTAMP, TypeSig.TIMESTAMP),
-          // interval support DAYTIME column or CALENDAR literal
-          ("interval", TypeSig.DAYTIME + TypeSig.lit(TypeEnum.CALENDAR)
-              .withPsNote(TypeEnum.CALENDAR, "month intervals are not supported"),
-              TypeSig.DAYTIME + TypeSig.CALENDAR)),
-        (timeAdd, conf, p, r) => new BinaryExprMeta[TimeAdd](timeAdd, conf, p, r) {
-          override def tagExprForGpu(): Unit = {
-            GpuOverrides.extractLit(timeAdd.interval).foreach { lit =>
-              lit.dataType match {
-                case CalendarIntervalType =>
-                  val intvl = lit.value.asInstanceOf[CalendarInterval]
-                  if (intvl.months != 0) {
-                    willNotWorkOnGpu("interval months isn't supported")
-                  }
-                case _: DayTimeIntervalType => // Supported
-              }
-            }
-          }
-
-          override def convertToGpu(lhs: Expression, rhs: Expression): GpuExpression =
-            GpuTimeAdd(lhs, rhs)
-        }),
-        GpuOverrides.expr[Abs](
-          "Absolute value",
-          ExprChecks.unaryProjectAndAstInputMatchesOutput(
-            TypeSig.implicitCastsAstTypes,
-            TypeSig.gpuNumeric + GpuTypeShims.additionalArithmeticSupportedTypes,
-            TypeSig.cpuNumeric + GpuTypeShims.additionalArithmeticSupportedTypes),
-          (a, conf, p, r) => new UnaryAstExprMeta[Abs](a, conf, p, r) {
-            val ansiEnabled = SQLConf.get.ansiEnabled
-
-            override def tagSelfForAst(): Unit = {
-              if (ansiEnabled && GpuAnsi.needBasicOpOverflowCheck(a.dataType)) {
-                willNotWorkInAst("AST unary minus does not support ANSI mode.")
-              }
-            }
-
-            // ANSI support for ABS was added in 3.2.0 SPARK-33275
-            override def convertToGpu(child: Expression): GpuExpression = GpuAbs(child, ansiEnabled)
-          }),
       GpuOverrides.expr[MultiplyYMInterval](
         "Year-month interval * number",
         ExprChecks.binaryProject(
@@ -198,17 +153,6 @@ trait Spark330PlusShims extends Spark321PlusShims with Spark320PlusNonDBShims {
           override def convertToGpu(lhs: Expression, rhs: Expression): GpuExpression =
             GpuMultiplyYMInterval(lhs, rhs)
         }),
-      GpuOverrides.expr[MultiplyDTInterval](
-        "Day-time interval * number",
-        ExprChecks.binaryProject(
-          TypeSig.DAYTIME,
-          TypeSig.DAYTIME,
-          ("lhs", TypeSig.DAYTIME, TypeSig.DAYTIME),
-          ("rhs", TypeSig.gpuNumeric - TypeSig.DECIMAL_128, TypeSig.gpuNumeric)),
-        (a, conf, p, r) => new BinaryExprMeta[MultiplyDTInterval](a, conf, p, r) {
-          override def convertToGpu(lhs: Expression, rhs: Expression): GpuExpression =
-            GpuMultiplyDTInterval(lhs, rhs)
-        }),
       GpuOverrides.expr[DivideYMInterval](
         "Year-month interval * operator",
         ExprChecks.binaryProject(
@@ -219,20 +163,9 @@ trait Spark330PlusShims extends Spark321PlusShims with Spark320PlusNonDBShims {
         (a, conf, p, r) => new BinaryExprMeta[DivideYMInterval](a, conf, p, r) {
           override def convertToGpu(lhs: Expression, rhs: Expression): GpuExpression =
             GpuDivideYMInterval(lhs, rhs)
-        }),
-      GpuOverrides.expr[DivideDTInterval](
-        "Day-time interval * operator",
-        ExprChecks.binaryProject(
-          TypeSig.DAYTIME,
-          TypeSig.DAYTIME,
-          ("lhs", TypeSig.DAYTIME, TypeSig.DAYTIME),
-          ("rhs", TypeSig.gpuNumeric - TypeSig.DECIMAL_128, TypeSig.gpuNumeric)),
-        (a, conf, p, r) => new BinaryExprMeta[DivideDTInterval](a, conf, p, r) {
-          override def convertToGpu(lhs: Expression, rhs: Expression): GpuExpression =
-            GpuDivideDTInterval(lhs, rhs)
         })
     ).map(r => (r.getClassFor.asSubclass(classOf[Expression]), r)).toMap
-    super.getExprs ++ map
+    super.getExprs ++ map ++ DayTimeIntervalShims.dayTimeIntervalExprs
   }
 
   // GPU support ANSI interval types from 330
