+++ sql-plugin/src/main/330db/scala/com/nvidia/spark/rapids/shims/SparkShims.scala
@@ -43,6 +43,9 @@ object SparkShimImpl extends Spark321PlusDBShims {
     new ParquetFilters(schema, pushDownDate, pushDownTimestamp, pushDownDecimal, pushDownStartWith,
       pushDownInFilterThreshold, caseSensitive, datetimeRebaseMode)
   }
+
+  override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] =
+    super.getExprs ++ DayTimeIntervalShims.dayTimeIntervalExprs
 }
 
 trait ShimExtractValue extends ExtractValue {
